// frontend/src/types/Image.ts
export interface ImageData {
  _id: string;
  url: string;
  title: string;
  author?: string;
  born_died?: string;
  date?: string;
  technique?: string;
  location?: string;
  form?: string;
  type?: string;
  school?: string;
  timeline?: string;
  description?: string;
  tags?: string[];
  uploadDate?: string;
  score?: number;
}